Updated dependencies to Rails 5.1. Updated dummy test app to Rails 5.1.
Updated docs to reflect changes in Rails 5.1 and how the test runner works differently to rake. Removed references to ApplicationController and Engines as this tool now based on Sinatra and not ApplicationController.

Updated docs with a section on development.

Removed the doc controller test as now running as Sinatra app.

Updated tests to consistently pass despite random order.

Updated the version, adding a sinatra flag to make distinct from source project.

Ran rubocop over the project to cleanse the syntax, targeted at ruby 2.4.0.